cricket-llm-bot/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
groq
